pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Josh K",email = "keisler@live.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "psycopg-binary (>=3.2.5,<4.0.0)",
    "psycopg (>=3.2.5,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "black (>=25.1.0,<26.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false